Handle the new package structure
Since RHEL 6.10 and RHEL 7.5 the package was renamed from
redhat-access-insights to insights-client. When systemd is present, a
timer unit is used instead of cron.

We use facts to automatically determine which package should be present,
but this can fail when content views are used. Because of that a new
parameter is introduced.